requirements.txt
accelerate==0.21.0
albumentations==0.4.3
bitsandbytes==0.41.0
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
datasets==2.8.0
deepspeed==0.9.5
diffusers==0.17.1
einops==0.6.1
einops-exts==0.0.4
fairscale==0.4.13
fastapi==0.101.1
flash-attn==2.0.7
gradio==3.35.2
gradio_client==0.2.9
httpx==0.24.0
huggingface-hub==0.16.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
invisible-watermark==0.1.5
k-diffusion @ git+https://github.com/crowsonkb/k-diffusion.git@c9fe758757e022f05ca5a53fa8fac28889e4f1cf
kornia==0.6.0
lpips==0.1.4
markdown2==2.4.10
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
omegaconf==2.1.1
openai==0.27.8
opencv-python==4.7.0.72
pandas==2.0.3
peft==0.4.0
Pillow==9.5.0
pudb==2019.2
pytorch-lightning==1.4.2
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentencepiece==0.1.99
shortuuid==1.0.11
streamlit==1.24.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
test-tube==0.7.5
timm==0.6.13
tokenizers==0.13.3
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
uvicorn==0.23.2
wandb==0.15.8
